Moved parser/lexers/tree walkers to net.cscott.sdr.calls.transform class, to keep the 'representation of calls and positions' code separate from the munging of the representation.
Split off TestParser into a separate class, instead of including a main
routine in CallFileParser() -- this also removes a circular dependency
between CallFileParser and CallFileBuilder which prevented the project
from being built cleanly from scratch.

Renamed *.g files to more closely match the names of the generated java
classes.

Started a Prim.warp() method to apply Warps to movement primitives.

Added BeatCounter/RemoveIn to re-time primitives in order to remove
'In' nodes from the Call representation.  Works on low-level asts,
with all the formation-selection and call-expansion already applied.

Added basic Walker class to walk the AST: idea is that it should be
easy to create derived AST walkers from this.