Extend parser to handle 'example' phrases in call definitions.
We prepend the ascii-art diagrams in the examples with exclamation marks
to make the lexer easier.  We also add 'ferris wheel' to the mainstream
definition list with an example of an example clause.

Partial support for 'ends with' and 'assert' clauses; I can't see exactly
where to fit them into the grammar at the moment, so they are not actually
parsed, even though we could lex them and have (unused) production rules
for them.

We don't currently store the examples in the call definition structure;
we should -- and we should have a test program that processes the
examples and checks them.