Tweak GeneralFormationMatcher so that resulting meta formation is symmetric.
Since we attempt to match dancers in order by sorted index (and our
symmetric-match detection is also based on ordering here), we can encourage
a symmetric result by ensuring that symmetric dancers in the input formation
are grouped together in the sort order.  Matching a miniwave against a wave
now results in another miniwave, instead of possibly resulting in a couple.