Add tag information to FormationMatch.toString().
Add tag information to FormationMatch pretty-printer.  We also added a
general dancer-ordering comparator to Formation, since we were reimplementing
it everywhere.  The FormationMatch pretty-printer sorts all output so that
it is not dependent on implementation-specific details such as the ordering
of keys within the FormationMatch.matches map, or the ordering of dancers
within the FormationMatch.meta formation.