Introduce NamedTaggedFormation.
Instead of using an anonymous inner class to associate nice names with
standard formations, factor out a "real" public NamedTaggedFormation
class, and use it.

Also fix the FormationList.main() self test code, and add some additional
doctests to FormationList that assert that canonical formations are centered
and have most of the dancers facing in the north/south direction.