Work on quarter tags in FormationList.
Add single (three-) quarter tags of various sorts, and use these to create the
8-person tags.  Apply proper tags to the formation (adding OUTSIDE_4, OUTSIDE_6,
and CENTER_6 tags) and fix orientation.

Also clean up some old commented-out code in FormationList.

In the future, we might need to add "ends out" and "diamond spots" variations
of the tags, as well as adding "tag spots" and "ends in" variations of the
diamonds.