Some convenience functions to make sets and maps more like first class objects.
Much of the convenience in writing in scripting languages like Python and Ruby
comes from the fact that lists, sets and maps are first class objects and
can easily be created and manipulated.  While full-scale list comprehensions
are probably out of easy reach for Java, we can at least make it much easier
to create simple lists, sets, pairs, and maps with a few elements.
(Pairs are a workaround for the fact that the map constructor wants to
alternate K and V type arguments.)

Use:
  import static net.cscott.sdr.util.Tools.*;
to make the l(), s(), p(), and m() constructors available.