Add new Evaluator framework, replacing the old Evaluate/Elaborate class.
Evaluation now occurs in a stack of Evaluators, each of which manages
its own continuation.  The result is accumulated in a mutable
DanceState object.  Concepts (implemented in the Call class) are now
split into two kinds: "simple" concepts (like fractional) which just
transform the subcalls in some manner, resulting in a call tree that
can be interpreted by the Standard Evaluator; and "complex" concepts
(like tandem) which use their own custom Evaluator class on their
children (often by doing some pre-processing, invoking the standard
evaluator, and then post-processing the results).  Transformations
which operate on the call tree (like Fractional) must now be careful
only to expand "simple" concepts.

The Standard Evaluator is finished, but it invokes a MetaEvaluator
to deal with 'FROM' clauses, and a PartsCombineEvaluator to deal with
'SELECT'; neither of this is complete yet.