Improvements to natural language grammar skeletons.
Turn off backtracking and up the lookahead, so that ANTLR tells us about
grammar ambiguities instead of suppressing those notices.  (It wouldn't be
an issue if ANTLR was fully backtracking, but in fact in commits to a
given rule alternative and won't backtrack out of multiple rules to try a
higher-level alternative.  This leads to subtle and non-obvious bugs.)

Tweak the grammar to get rid of ambiguities and add "typing" shortcuts
(digits for typed-out-english-phrases) in a way that won't cause parsing
problems.  Loosen the fraction rule (accept "do half of a scoot back") but
tighten the cardinal rule ("scoot back 1 time" is out; say, "scoot back once").

Change the grammar start production to make sure we're attempting to match
the EOF at the end of the phrase.  Add some doc tests showing how we resolve
some grammar ambiguities.  (These tests don't all succeed yet.)