Add SASHAY_START and SASHAY_FINISH flags to Prim and DancerPath.
Movements such as half-sashay and dodge involve motions which do not
begin in the "step forward" direction.  Add bits to DancerPath and
Prim to flag these, so that we don't animate the motion as an S-shape
beginning and ending in the facing direction.

Update lots of doctests to match, since the new flags show up in the
output from tests involving dosados.