Emit post-processed version of natural language grammar as Java source.
We can use the post-processed version as input to a call completion engine,
without having to re-run the processing steps on every startup.  (The
post-processed version handles precedence and left recursion.)

We might consider using this same trick on the CallDB and FormationList at
some point, to avoid having to parse and load those everytime.