Clean up natural language grammar.
Don't allow "scoot back once" -- it has to be more than 1 repetition.
Allow "do half of a scoot back" but not "scoot back once and half": be
more careful about the context in which the article can be elided.
Use a separate <NUMBER> token in the java grammar, instead of trying
to parse digits as part of the <number> nonterminal.
Finally, allow "two quarters" as a synonym for "half", because sometimes
you will call "swing and circle two quarters", for example.