Rework evaluation of In.
We resurrect the 'RemoveIn' class and rework it a bit to make it actually
useful (instead of interesting but dead code).  We avoid pushing In nodes
down 'too far' so we don't do unnecessary work on the 'wrong side' of a
If or Opt -- but be careful when writing recursive definitions, because the
BeatCounter does recurse.  Careful placement of In nodes (so that recursion
is not necessary) avoids this: eg, instead of putting a "2 beats per hand"
In at the top of the square thru N definition (which would require recursion
to count and then evenly distribute the beats) we push the Ins down in the
definition so that each part has an appropriate number of beats.  Since
recursive call definitions are rare, we don't expect this to be a problem
in practice.

We also add an '_in' pseudo-concept, so that we can push In nodes into
Apply nodes without expanding them, and rewrite the definition of
'square thru' to use this concept.