Allow "do half of a fan the top".
Add an intermediate halfway point for the fan the top definition.  This
improves the animation, as the breathing now happens in two stages instead
of all at once at the end.

In order to make this work, I also had to extend the call file grammar to
accept 1/8 rotations, and the Ast grammar to accept negative fractions as
rotation amounts.