Error reporting improvements.
Split formation name and reason in the NoMatchException constructor, so we
can construct meaningful messages for compound NoMatchExceptions from ORs and
the line.  Name formations joined by OR for this purpose as well.

Improve messages generated by Evaluators and OR selectors (and update doctests!)

Use NamedTaggedFormation for _ends_in to make sure we don't lose the
selector name!  Also use NamedTaggedFormation for ANY and NONE selectors.