PMSD: support nested invocations sharing the same state object.
We also add an 'errorDetails' boolean property to the state object.  The idea
here is that we can set /errorDetails=true and then /runTest('foobar') and
get additional information about the failing part of the test.  Perhaps
other uses can be found as well -- we can turn on/off parts of the test
with something like
   /if (someVariable) exit;
for example.