Rework predicate evaluation to allow returning non-boolean types.
Previously, something like EQUAL(ADD(1,1),2) could not be evaluated, because
there was no way for the ADD(1,1) part to return something other than a boolean.
Now we have evaluateAsString/evaluateAsNumber, and we parse:
  EQUAL(ADD(1,1), 2)
as:
  (Condition equal (Condition add (Condition literal (Condition 1))
                                  (Condition literal (Condition 1)))
                   (Condition literal (Condition 2)))
So the only time that the 'predicate' field is interpreted literally as a
string is when the 'literal' predicate evaluates its single argument.

In the callfile parser, we treat 'num-dancers' as
   (Condition literal (Condition num-dancers))
but 'num-dancers()' as
   (Condition equal)
This lets us define no-argument predicates that return a non-boolean value.