Breathing work: remove unnecessary breathing in 'from' evaluation.
We only breathe the input formation to a 'from' -- breathing the input
ensures that we don't pile up collisions to the point where we can't
resolve them later, and not breathing the output ensures that we don't
unnecessarily screw with space invaders and 'do your parts'.  (Note that
'input' and 'output' may be reversed from your initial idea here, the
'input' is the meta formation, and the 'output' are the bits we're
substituting into the meta.)

We still have a breather at top-level -- we'll have to fix that eventually
to make calls like "press ahead" work.  (Formations with phantoms might
also need to avoid breathing -- or they might just instantiate phantoms
at the top level so that the breathing works out.)  Probably we'll need
to set a 'space invader' flag on the formation or in the dance state
to inhibit the top-level breather when necessary.