Fix breathing of 1/2 run situation.
This was complicated: because we try to resolve unshared edges first, the
completely-overlapped X axis of the end dancers (who share fewer edges)
was trying to be trimmed (unsuccessfully) before the shared-but-less-overlapped
Y axis trim bits.  We now put all completely-overlapped trim bits at the end
of the list... but this feels like a hack to me; I wouldn't be surprised if
I have to revisit this again.  But we've got a test case at least to ensure
that we don't regress if I end up having to fix this another way.