Rewrite core of breathing algorithm to use Cassowary linear constraint solver.
Replace our ad-hoc multi-pass constraint resolver with a "real" one based on
the LGPL Cassowary library (reworked to use exact rational arithmetic).  This
makes the behavior of the algorithm much more obvious and eliminates problems
with unusual formations (like the 'half of an ends cross run' in the
cross-run-1 test).  It may be worth revisiting the trimOverlaps() routine
in the breather to use the constraint solver, too.