Extend Cassowary to handle mixed integer linear programming problems.
We use the branch-and-bound algorithm, invoking Cassowary to solve subproblems,
in order to handle mixed integer linear programming problems.  This lets us
introduce boolean variables to handle overlap constraints ("either the x
dimension should not overlap, or the y dimension should not overlap").
Extend BreatheTest to include a proof-of-concept of this method of resolving
formation overlaps.