Tweaks to ClBranchAndBound.
Decrease the LARGE_NUMBER to 100 to avoid overflow problems.

Check for ClEditOrStayConstraints, because the way that stay constraints
behave when constraints are removed makes our branch-and-bound algorithm
misbehave.  (Stay constraints are reset to their "current values" when a
constraint is removed; we need them to 'stay' where they were when we
began to solve() so that we can continue evaluating other branches.)