Add default arguments to the call list grammar; more <anyone> <anything> work.
Allowing arguments to have default values helps us work around a grammar
infelicity: because ANTLR doesn't support left recursion, we have to combine
the "<anyone> <anything>" and "<anyone> <anything> while the others <anything>"
into one rule, with an optional second part.  That means we need to specify
what the second <anything> defaults to ("nothing"), if it is missing.

We could go back and use this mechanism to let "square thru" default to
"square thru(4)", etc.  Not going to do that in this commit, but maybe in
the future.