Lazy evaluation of predicates with substitution and arguments.
We now represent the nested calls of the arguments in the condition node,
instead of eagerly evaluating it, which also allows us to better match
calls with arguments in the CALL IS predicate.

We might have to eventually have a 'force' call which eagerly evaluates its
argument, since we don't memo-ize arithmetic results.  But we don't have
much math, really, so unmemoized lazy evaluation should work fine for now.