Fix bogus ExprFunc<Object> type for LITERAL.
The only instance of Class<? super Object> is Object.class, which
isn't really what we meant for LITERAL to take.  We now use an erased
type.  We could have made a LITERAL() method in the style of
Collections.emptyList() which returns the ExprFunc<T> of your choice,
but the erased type seems cleaner.