Work on predicates: add string matching, regularize naming.
We have three varieties of 'equals' predicates now, named according
to the type of their arguments: EQUAL STR, EQUAL NUM, EQUAL CALL
(this last does a tree comparison).  (Replaces EQUAL/CALL IS)

Introduce a regexp matcher for strings (MATCH) and several functions which
generate strings from DanceStates: _SELECTION PATTERN, _ROLL PATTERN, and
_SWEEP PATTERN (Replaces the old _SELECTION PATTERN call which did both
the string-generation and the matching part).

To allow regexps as arguments for MATCH, added quoted-string support
to both the CallFile and Ast parsers.  We support java-style backslash
escapes in the string, too.  (This replaced the older quoted-string
special case for the message after a condition node.)

Regularize the name of the mathematic operators: we use spaces to separate
words, not underscores. (_MULTIPLY_NUM etc -> _MULTIPLY NUM etc)

Add a LESS predicate to go along with the GREATER predicate.