Define PRESERVE_ROLL/PRESERVE_SWEEP prim flags; make "and roll" preserve sweep.
We removed the default "standing still preserves roll" behavior and
replaced it with an explicit PRESERVE_ROLL flag where that's the
desired behavior.  We also removed the code which cleared sweep unless the
dancers ended facing in, since in recycle/wheel and deal/etc we need to
have the sweep there so that the final "and roll" can preserve it.  The
FACING COUPLES restriction on sweep will make sure the dancers are facing
the appropriate direction to sweep.  (Eventually, we'll also need to make
sure that the actual centers points match, to avoid "single wheel and
sweep 1/4" from a RH wave being a legal usage; see 'sweep-3' test case.)

We also needed to tweak the Ast grammar to allow dashes in flag names.

We also adjusted a lot of doc tests to add the new roll/sweep/preserve flags.