Add post-condition 'ends in' statement.
We add a new field to If in the AST, indicating whether the condition should
be evaluated before or after its child.  This allows us to define
post-conditions: try 'this', and then check that 'that' is true.

With a new 'formation' predicate, this lets us define assertions on the
ending formation of a call.