Add a selector function to match against a new formation.
Usually Selectors match against the current TaggedFormation (created by
a Matcher).  The 'formation' selector function lets you apply a new
Matcher to the current dance state and select dancers based on the tags
applied, without changing the current dance state.  This is primarily for
the use of post-conditions, which work against an untagged result formation,
but it will probably have other uses as well.