Add CIRCLE_ADJUST matcher for circle choreography.
This switches to the "other grouping" of dancers in a circle -- ie,
"dancer and corner" instead of "dancer and partner".  Since this adjusts
facing direction by 1/8, call definitions need to make sure they don't
turn more than 1/4 so that the total turn is less than 1/2 (and thus the
animation routines can unambiguously determine the direction of rotation).