Fix grammar precedence problems.
The new "heads down the middle and <anything>" call was conflicting
with the "and roll" and "and sweep" rules, since "roll" and "sweep"
are <anything>s.  Adjusted some precedence levels to fix the problem
(and fixed a bug to allow specifying precedence of calls which don't
take arguments).

Adjusted "<anyone> <anything>", "down the middle <anything>", "as couples",
"tandem", "siamese", and "roll".  Added "and individually roll" as a
looser-binding version of roll.  "tandem trade and roll" means roll in tandem;
"tandem trade and individually roll" means only apply tandem to the trade.

I think the precedence levels make sense after this commit.