Add "you're home" call.
This just checks that the dancers are actually home.

Tweaked grammar to allow apostrophes in spoken text, and to make
substitution in the emitter backslash-safe.

Added _COUPLE_NUM_PATTERN function, to aid in checking whether the
dancers are home.

Wrote some notes about how this might work for a singing call progression.