Add example from Clark's "Beat Math" paper; test that timings are correct.
Added a few tweaks to the call definitions and timings to make this work.
At the moment we *don't* use callerlab's convention, that squared set calls
ends in another squared set; we use the "challenge dancer's convention"
instead which ends in a 2x4.  Then we adjust on some of the old-timey calls
to allow them to adjust to squared set spots from a 2x4 where needed.