Expose environment to the dance engine via a property map in DanceState.
First use of this is via:
  condition: MATCH(_PROPERTY(call-pending), true)
for calls like circle left, which should "keep circling until the next call
is given".   We can add additional properties to the map if we wish to expose
other environment state.  By using a ConcurrentMap, the environment can
safely interact asynchronously with the dance engine.