Make speech interrupter more aggressive about pulling data.
This ensures that interruptions occur at the proper place in the data stream,
even if the recognizer is running behind real time.  It do anything about
excessive buffering by the input device (such as happens with USB
microphones), but at least it ensures the problem isn't compounded by
the recognizer.