Fix bug in speechInterrupter which blocked data after an interruption.
The "all empty properties" packet we were constructing was giving the
RaisedCosineWindower a fit -- the data packet had -1s for sample rate,
which was being interpreted as a large positive number and hanging the
windowing process while it computed cosine transformations over a
very large window indeed.  Anyway: just use the first packet we see as
our pad packet for the future.  That's much nicer to the rest of the
pipeline.