Basic implementation of "initially" concept (C3A).
Had to hack in a "concept" nonterminal in the grammar.  Should really
extract this automagically from the definitions somehow.  Refactored
"finally" Call object to make it easier to add the rest of the
family of similar concepts ("like a", "_first part", etc).

Simple concepts currently work, but not complicated concepts requiring
currying.