Parse 'optional' clauses for CONCEPT/SUPERCALL/METACONCEPT.
Automatically extract concepts/supercalls/metaconcepts from the
call file definitions when they are tagged via the 'optional' clause
(which should probably be renamed to the 'option' clause).
Handle left/reverse the same way -- in BuildGrammars, not as a hack
in CallFileBuilder.g.

Added a bogus VOID token which will never be emitted by the lexer
in order to tie together the lower-level grammars when there are no
concepts/metaconcepts defined.  Slightly hacked ANTLR to suppress
an 'unused token' warning for VOID, since it was noisy during builds.
New hacked source file is included in the antlr .jar.