Add `condition` as a Matcher function.
This lets you embed predicates inside `from:` clauses.  This will
be useful when I add else/commit capability to from, so that we
don't try other clauses once a given `from:` matches.  This will
let us simplify a number of definitions where we need to restate
complementary versions of the same condition for flow control.