Improve `Formation.mapStd()` and `setFormation` in PMSD.
Add a `normalCouples` field and a `mapStd()` override to
`NamedTaggedFormation` so that we can ensure that we have normal
couples for well-known formations in `FormationList`.  Update
the default `setFormation` command in PMSD to use `mapStd` if the
given formation doesn't have `StandardDancer`s.  This makes it
much easier to write test cases in PMSD from arbitrary starting
formations.