Handle "star like" outside formations in ConcentricEvaluator.
Concentric Evaluator couldn't keep the centers and ends on 1x2 spots
when evaluating 'four ladies chain' because the ends breathed in to
a formation with a 1x1 hole in the center.  Handle this situation
and move in the ends where needed.

Rework definition of 'four ladies chain' slightly to enforce the
star promenade formation constraints that are needed to make it work.
(The old _finish belles chain only worked if the starting positions
were star promenade spots -- but the concentric evaluator was
actually placing the ends further out, which was causing issues.)