Add Rotation.intersect(), Rotation.consistent(), Position.relocate(Rotation).
When doing matching against multiple inexact rotations, what we really want
to know is if they are "consistent" -- is there some rotation in the
intersection of both of their congruence sets.

Position.relocate(Rotation) is a convenience for a common relocation, where
we are just replacing an inexact rotation with an exact one or vice-versa.