EvalPrim.apply() can now dance with phantoms.
We can't always apply a Prim to a dancer with an inexact rotation, but
if the direction is to stand still or is given using in/out/roll/sweep
modifiers we often can.

Dancer paths can now have inexact facing directions; thankfully these
paths are removed before we actually have to animate them.