Add FormationMatch.inserted to track phantom dancers.
When evaluating "triple lines" or other formations which insert phantoms,
we need to keep track of which phantoms we've added so that we can remove
them again after the call is evaluated in the formation.  There may be
multiple nested concepts which each add their own phantoms, so we can't
just remove all NONCORPOREAL dancers, we need to be specific about which
formation match added which.