Implement '_o concentric' concept.
This leads the ends as they are, so calls like 'settle back' with an
'ends o circulate twice' can be done.  We still need to breathe out
the ends occasionally though; consider "1/2 bits and pieces" from
columns.  We also implement distorted concepts like Os by removing the
distortion, doing the call, then reapplying the distortion -- so
'_o concentric' needs to know how to handle the ends intruding into
the centers' space as well.