Path blending: _blend first, _blend last.
We use path blending to animate many "in one smooth motion" calls and
distortion concepts.  "Follow your neighbor and spread" can blend the
spread with the final part of follow your neighbor.  The "O" concept
can compress the O, do the call in a column, and then expand the O,
blending the first and last steps the way real dancers would.  Several
"cross" calls can be defined using a blended away spread or sashay;
for example, cross extend can be defined as a blended extend and slide.

One current limitation is that we always blend the "last motion", but
sometimes that doesn't actually correspond to the "last part" because
there are intermediate positions defined: for example, circulate and
spread both have "1/2 circulate" and "1/2 spread" intermediate positions.
We kludge around this for now, but we should eventually tweak the
blend concepts to take an optional argument specifying "how much to
blend".