Improve auto-tagging of 8-person formations; quarter tags in particular.
The GENERAL_QUARTER_TAG formation was missing its VERY_CENTER tags,
and the other tag formations were missing the OUTSIDE_2 tags.  Fixed
this problem in general by improving the autotagger to apply the various
center/outside 2/4/6 tags on 8-person formations.  While we're at it,
extend the autotagger to work on 12-person and 16-person formations.

There is still some bug in the 2-person autotagger, and the quarter tag
ends don't get their BEAU/BELLE tags applied.  Do this manually for now.

Banish OUTSIDE_4 for everything except tidal waves, where it awkwardly
describes how the "centers" of a tidal wave are conventionally the
centers of each 1x4.  This is a quirk that should probably be refactored
away and dealt with as a special case in the "centers(x)" definition.
But for now, OUTSIDE_4 lives on for this one special case.  Everything
else uses CENTER/END to mean CENTER_4/OUTSIDE_4.

Fix some tagging bugs in butterfly/O formations (where the tags should
match those in the undistorted column) and 4x4 formations (where the
"center 8" is ambiguous unless you specify lines or columns).