Ensure optimal solution for insideBox(Formation).
Consider an initial bound of (0,0;5,3) and dancers at (1,2;2,3) and (2,1;3,2).
If we consider the dancer at (2,1;3,2) first we will decide to trim on X
because (0,0;2,3) is area 6 while (0,0;5,1) is area 5.  But once you
factor in the other dancer, the area 5 region is actually optimal.

Consider all possible trimmings to ensure that the box returned by
insideBox does actually have the maximum possible area.