Refactor chain reaction.
Create `_turn the diamond` and `_turn the star` helpers.  The former tries
to maintain the initial elongation of the diamond, even if the turn is 1/4
or 3/4.  The latter smashes the diamond into a star first, which is more
appropriate for "those who meet" calls.  (See `_unsmash the star` for how
to pull out those who meet.)