Make 'transfer and' work for non-2x4 formations.
Two things necessary: first, add new `TRANS COLUMN` formations/matchers which
ensure that #1 and #2 are in the correct place for a transfer, ignoring the
other dancers.  Use these in selectors as well.  Be careful not to
actually try to use `TRANS COLUMN` directly in a `from` clause, as the
resulting breathing of unmatched dancers is probably not what you want.

Second, add a new `Selector` parameter to the various `ConcentricEvaluator`s
which gives more control over who we consider to be the centers.  This lets
us use `_with designated` to tag the #1 and #2 dancers at the start of the
transfer, and then use `not(DESIGNATED)` to pick out the centers later.
This fixes problems with "trans grand chain eight" where the "centers" were
actually further out that the #1/#2 tandem after breathing.